% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartRacer.R
\name{add_gg_run}
\alias{add_gg_run}
\title{Add a workout to map}
\usage{
add_gg_run(gmap, rcd, size = 1, pch = 1)
}
\arguments{
\item{gmap}{ggmap object}

\item{rcd}{A record from}
}
\description{
Add a workout to map
}
