% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartRacer.R
\name{heartRace}
\alias{heartRace}
\title{Animate fit files onto google maps}
\usage{
heartRace(recordList, zoom = 14, map_type = "satellite", nudgeLatBy = 0,
  nudgeLonBy = 0, timeName = "timestamp", latName = "position_lat",
  lonName = "position_long", hrName = "heart_rate")
}
\arguments{
\item{recordList}{A list of records from fit file}

\item{zoom}{Zoom option for google maps}

\item{map_type}{Type of map pulled from google maps}

\item{nudgeLatBy}{How much should each subject's latitude be moved by?}

\item{nudgeLonBy}{How much should each subject's longitude be moved by?}

\item{timeName}{Column name of timestamp in record}

\item{latName}{Column name of latitude in record}

\item{lonName}{Column name of longitude in record}

\item{hrName}{Column name of heart rate in record}
}
\description{
Animate fit files onto google maps
}
\details{
Returns a ggplot object with the work out plotted on top of it. This can be animated
by using \code{gganimate::gganimate(hrMap)}, where \code{hrMap} is an object returned by heartRace.
}
