% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartRacer.R
\name{race_points}
\alias{race_points}
\title{Animate points for group of subjects}
\usage{
race_points(x, y, time, hrtRt, id, speed_up = 120, ptsPerSec = 3,
  ani = TRUE, lgdLocation = NULL, googMap = NULL, ggmap = NULL, ...)
}
\arguments{
\item{x}{Longitude}

\item{y}{Latitude}

\item{time}{Time is seconds. Should be minimized to 0}

\item{hrtRt}{Heart Rate}

\item{id}{Id correpsonding to other variables}

\item{speed_up}{How much faster should animation run than in reality?}

\item{ptsPerSec}{How many points plotted per second?}

\item{ani}{Is this an animation?}

\item{lgdLocation}{Where onshould the legend be plotted?}

\item{googMap}{Google map object to be plotted on}

\item{...}{arguments that will be ignored. Only for compatibility with \code{RgoogleMaps::plotmap}}
}
\description{
Animate points for group of subjects
}
