% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartRacer.R
\name{upload_fit}
\alias{upload_fit}
\title{Upload fit files from device}
\usage{
upload_fit(device_location = "~/../../Volumes/GARMIN/GARMIN/ACTIVITY",
  destination = "~/Documents/fitFiles")
}
\arguments{
\item{device_location}{Location of device}

\item{destination}{Where saved files should go}
}
\description{
Upload fit files from device
}
\details{
Grabs new fit files from device and saves them locally
}
