% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartRacer.R
\name{heartRace}
\alias{heartRace}
\title{Awesome function for being awesome}
\usage{
heartRace(recordList, zoom = 14, use_rgoogs = F, use_ggmap = T,
  map_type = "terrain", nudgeLatBy = 0, nudgeLonBy = 0)
}
\description{
Awesome function for being awesome
}
